backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
pydantic-core==2.20.1
python-multipart==0.0.9
httpx==0.27.2
openai==1.51.0
slowapi==0.1.9
PyYAML==6.0.2
pytest==8.3.2
pytest-asyncio==0.23.8
anyio==4.4.0